+++ src/components/InRequest/InRequestEditPanel.tsx
@@ -19,6 +19,7 @@ import {
   Combobox,
   OptionGroup,
   Option,
+  InfoLabel,
 } from "@fluentui/react-components";
 import { DatePicker } from "@fluentui/react";
 import { PeoplePicker } from "components/PeoplePicker/PeoplePicker";
@@ -101,8 +102,7 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
   const currentUser = useContext(UserContext).user;
 
   // Create a type to handle the IInRequest type within React Hook Form (RHF)
-  type IRHFInRequest = Omit<IInRequest, "MPCN"> & {
-    MPCN?: string;
+  type IRHFInRequest = IInRequest & {
     SSN?: string;
   };
 
@@ -117,7 +117,7 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
     criteriaMode:
       "all" /* Pass back multiple errors, so we can prioritize which one(s) to show */,
     mode: "onChange" /* Provide input directly as they input, so if entering bad data (eg letter in MPCN) it will let them know */,
-    values: { ...props.data, MPCN: props.data.MPCN?.toString() },
+    values: { ...props.data },
   });
   const updateRequest = useUpdateRequest(props.data.Id);
   const additionalInfo = useAdditionalInfo(props.data.Id);
@@ -129,6 +129,37 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
   const eta = watch("eta");
   const employee = watch("employee");
   const workLocation = watch("workLocation");
+
+  const validateMPCN = (value?: string) => {
+    if (!value || value === "") {
+      return "MPCN must be at least 7 characters";
+    } else if (
+      value.length !== 0 &&
+      value.length < 7 &&
+      !"RAND000-".startsWith(value.toUpperCase()) &&
+      !value.match(/^\d{0,6}$/)
+    ) {
+      return "MPCN cannot contain non-numeric characters in the first 6 positions, unless it starts with 'RAND000-'";
+    } else if (
+      value.length !== 0 &&
+      value.length < 8 &&
+      "RAND000-".startsWith(value.toUpperCase())
+    ) {
+      return "MPCNs starting with 'RAND000-' must be followed by 6 digits";
+    } else if (value.length < 7) {
+      return "MPCN must be at least 7 characters";
+    } else if (
+      value.length > 7 &&
+      !value.toUpperCase().startsWith("RAND000-")
+    ) {
+      return "MPCN cannot be more than 7 characters, unless it starts with 'RAND000-'";
+    } else if (value.length > 7 && !value.match(/^RAND000-\d{6}$/i)) {
+      return "MPCNs starting with 'RAND000-' must be followed by 6 digits";
+    } else if (value.length === 7 && !value.match(/^\d{6}[A-Z|a-z|0-9]$/i)) {
+      return "MPCNs that are 7 characters must either be 7 digits, or 6 digits followed by a letter";
+    }
+    return;
+  };
 
   const gradeRankOptions = useMemo(() => {
     switch (props.data.empType) {
@@ -158,19 +189,7 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
   }, [eta]);
 
   const updateThisRequest = (data: IRHFInRequest) => {
-    // Translate the string MPCN to an Integer
-    let mpcn: number | undefined;
-    if (
-      data.empType === EMPTYPES.Civilian ||
-      data.empType === EMPTYPES.Military
-    ) {
-      // Parsing the string should be fine, as we enforce pattern of numeric
-      if (data.MPCN) {
-        mpcn = parseInt(data.MPCN);
-      }
-    }
-
-    const data2 = { ...data, MPCN: mpcn, SSN: undefined } as IInRequest;
+    const data2 = { ...data, SSN: undefined } as IInRequest;
 
     // Is the SSN different than what came in and therefore needs pushed to the server
     if (
@@ -626,10 +645,26 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
               )}
             </div>
             <div className={classes.fieldContainer}>
-              <Label
+              <InfoLabel
                 htmlFor="MPCNId"
                 size="small"
                 weight="semibold"
+                info={
+                  <Text>
+                    Acceptable formats are:
+                    <ul>
+                      <li>
+                        <b>Standard:</b> 7 digits
+                      </li>
+                      <li>
+                        <b>Over hire:</b> 6 digits + 1 letter
+                      </li>
+                      <li>
+                        <b>PAQ/COPPER CAP:</b> 'RAND000-' + 6 digits
+                      </li>
+                    </ul>
+                  </Text>
+                }
                 className={classes.fieldLabel}
                 required={
                   props.data.empType === EMPTYPES.Civilian ||
@@ -638,28 +673,12 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
               >
                 <NumberFieldIcon className={classes.fieldIcon} />
                 MPCN
-              </Label>
+              </InfoLabel>
               <Controller
                 name="MPCN"
                 control={control}
                 rules={{
-                  required:
-                    (props.data.empType === EMPTYPES.Civilian ||
-                      props.data.empType === EMPTYPES.Military) &&
-                    "MPCN is required",
-                  minLength: {
-                    value: 7,
-                    message: "MPCN cannot be less than 7 digits",
-                  },
-                  maxLength: {
-                    value: 7,
-                    message: "MPCN cannot be more than 7 digits",
-                  },
-                  pattern: {
-                    value:
-                      /^\d+$/ /* We don't want the pattern to enforce 7 numbers so we can have a unique error for non-numeric (eg letters/symbols) */,
-                    message: "MPCN can only consist of numbers",
-                  },
+                  validate: validateMPCN,
                 }}
                 render={({ field }) => (
                   <Input
@@ -670,7 +689,9 @@ export const InRequestEditPanel: FunctionComponent<IInRequestEditPanel> = (
                     }
                     disabled={props.data.empType === EMPTYPES.Contractor}
                     id="MPCNId"
-                    inputMode="numeric"
+                    onChange={(_event, data) =>
+                      field.onChange(data.value.toUpperCase())
+                    }
                   />
                 )}
               />