+++ src/components/InRequest/InRequestNewForm.tsx
@@ -20,6 +20,7 @@ import {
   Combobox,
   Option,
   OptionGroup,
+  InfoLabel,
 } from "@fluentui/react-components";
 import { IInRequest, useAddRequest } from "api/RequestApi";
 import { useForm, Controller } from "react-hook-form";
@@ -81,11 +82,10 @@ const useStyles = makeStyles({
 });
 
 // Create a type to handle the IInRequest type within React Hook Form (RHF)
-type IRHFInRequest = Omit<IInRequest, "empType" | "workLocation" | "MPCN"> & {
+type IRHFInRequest = Omit<IInRequest, "empType" | "workLocation"> & {
   /* Allowthese to be "" so that RHF can set as Controlled rather than Uncontrolled that becomes Controlled */
   empType: EMPTYPES | "";
   workLocation: worklocation | "";
-  MPCN: string;
   SSN: string;
 };
 
@@ -117,6 +117,37 @@ const InRequestNewForm = () => {
   const employee = watch("employee");
   const workLocation = watch("workLocation");
   const SAR = watch("SAR");
+
+  const validateMPCN = (value?: string) => {
+    if (!value || value === "") {
+      return "MPCN must be at least 7 characters";
+    } else if (
+      value.length !== 0 &&
+      value.length < 7 &&
+      !"RAND000-".startsWith(value.toUpperCase()) &&
+      !value.match(/^\d{0,6}$/)
+    ) {
+      return "MPCN cannot contain non-numeric characters in the first 6 positions, unless it starts with 'RAND000-'";
+    } else if (
+      value.length !== 0 &&
+      value.length < 8 &&
+      "RAND000-".startsWith(value.toUpperCase())
+    ) {
+      return "MPCNs starting with 'RAND000-' must be followed by 6 digits";
+    } else if (value.length < 7) {
+      return "MPCN must be at least 7 characters";
+    } else if (
+      value.length > 7 &&
+      !value.toUpperCase().startsWith("RAND000-")
+    ) {
+      return "MPCN cannot be more than 7 characters, unless it starts with 'RAND000-'";
+    } else if (value.length > 7 && !value.match(/^RAND000-\d{6}$/i)) {
+      return "MPCNs starting with 'RAND000-' must be followed by 6 digits";
+    } else if (value.length === 7 && !value.match(/^\d{6}[A-Z|a-z|0-9]$/i)) {
+      return "MPCNs that are 7 characters must either be 7 digits, or 6 digits followed by a letter";
+    }
+    return;
+  };
 
   const gradeRankOptions = useMemo(() => {
     switch (empType) {
@@ -150,19 +181,8 @@ const InRequestNewForm = () => {
   }
 
   const createNewRequest = async (data: IRHFInRequest) => {
-    // Translate the string MPCN to an Integer
-    let mpcn: number | undefined;
-    if (
-      data.empType === EMPTYPES.Civilian ||
-      data.empType === EMPTYPES.Military
-    ) {
-      // Parsing the string should be fine, as we enforce pattern of numeric
-      mpcn = parseInt(data.MPCN);
-    }
-
     const data2 = {
       ...data,
-      MPCN: mpcn,
       reqType: "In",
       SSN: undefined,
     } as IInRequest;
@@ -577,10 +597,26 @@ const InRequestNewForm = () => {
         )}
       </div>
       <div className={classes.fieldContainer}>
-        <Label
+        <InfoLabel
           htmlFor="MPCNId"
           size="small"
           weight="semibold"
+          info={
+            <Text>
+              Acceptable formats are:
+              <ul>
+                <li>
+                  <b>Standard:</b> 7 digits
+                </li>
+                <li>
+                  <b>Over hire:</b> 6 digits + 1 letter
+                </li>
+                <li>
+                  <b>PAQ/COPPER CAP:</b> 'RAND000-' + 6 digits
+                </li>
+              </ul>
+            </Text>
+          }
           className={classes.fieldLabel}
           required={
             empType === EMPTYPES.Civilian || empType === EMPTYPES.Military
@@ -588,29 +624,13 @@ const InRequestNewForm = () => {
         >
           <NumberFieldIcon className={classes.fieldIcon} />
           MPCN
-        </Label>
+        </InfoLabel>
         <Controller
           name="MPCN"
           control={control}
           defaultValue={""}
           rules={{
-            required:
-              (empType === EMPTYPES.Civilian ||
-                empType === EMPTYPES.Military) &&
-              "MPCN is required",
-            minLength: {
-              value: 7,
-              message: "MPCN cannot be less than 7 digits",
-            },
-            maxLength: {
-              value: 7,
-              message: "MPCN cannot be more than 7 digits",
-            },
-            pattern: {
-              value:
-                /^\d+$/ /* We don't want the pattern to enforce 7 numbers so we can have a unique error for non-numeric (eg letters/symbols) */,
-              message: "MPCN can only consist of numbers",
-            },
+            validate: validateMPCN,
           }}
           render={({ field }) => (
             <Input
@@ -618,8 +638,10 @@ const InRequestNewForm = () => {
               disabled={empType === EMPTYPES.Contractor}
               aria-describedby="MPCNErr"
               id="MPCNId"
-              inputMode="numeric"
               placeholder={empType === EMPTYPES.Contractor ? "N/A" : ""}
+              onChange={(_event, data) =>
+                field.onChange(data.value.toUpperCase())
+              }
             />
           )}
         />