+++ CHANGELOG.md
@@ -10,11 +10,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - (Keep your changes here until you have a release version)
 
+## [2.0.5] - 2023-11-16
+
 ### Changed
 
 - Removal of DD2875 process from In Processing
 - Added Job/Duty Title to In Processing
 - Added Duty Phone # to In Processing
+- Update MPCN to accept the following formats
+    a) Standard - 7 digits
+    b) Over hire - 6 digits + 1 letter
+    c) PAQ/CopperCap - "RAND000-" + 6 digits
 
 ## [2.0.4] - 2023-09-28